Moving transaction stats to stats module
