+++ src/stats/stats.controller.ts
@@ -1,11 +1,12 @@
 import { Controller, Req, Res, Get } from '@nestjs/common';
-import { ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ApiOperation, ApiParam, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
 
 import { Response, Request } from 'express';
 
 import { StatsService } from './stats.service';
 import { SupplyService } from './supply/supply.service';
 import { OperationsService } from './operations/operations.service';
+import { TransactionService } from '../transaction/transaction.service';
 
 @Controller('stats')
 @ApiTags('stats')
@@ -14,6 +15,7 @@ export class StatsController {
     private readonly service: StatsService,
     private readonly operations: OperationsService,
     private readonly supply: SupplyService,
+    private readonly transactions: TransactionService,
   ) { }
 
   @Get('/operations')
@@ -72,5 +74,50 @@ export class StatsController {
       return res.status(500).json({ error: `failed to get max supply: ${error}` });
     }
 
+  }
+
+  @Get('/transactions/:type/:from/:to')
+  @ApiOperation({ summary: 'Get transaction count per day' })
+  @ApiParam({ name: 'type', description: 'Transaction type'})
+  @ApiParam({ name: 'from', description: 'Date as `year-month-day` or timestamp in ms' })
+  @ApiParam({ name: 'to', description: 'Date as `year-month-day` or timestamp in ms' })
+  @ApiResponse({ status: 200 })
+  @ApiResponse({
+    status: 400,
+    description: ['invalid from date given', 'invalid to date given', 'invalid type given',
+      'invalid period range given'].join('<br>'),
+  })
+  @ApiResponse({ status: 500, description: `failed to get transaction stats '[reason]'` })
+  async getTransactionStats(@Req() req: Request, @Res() res: Response): Promise<Response> {
+    const fromParam = req.params.from;
+    const toParam = req.params.to;
+
+    let from = 0;
+    let to = 0;
+
+    from = Math.floor(new Date(fromParam.match(/\D/) ? fromParam : Number(fromParam)).getTime() / 86400000);
+    to = Math.floor(new Date(toParam.match(/\D/) ? toParam : Number(toParam)).getTime() / 86400000);
+
+    if (Number.isNaN(from)) {
+      return res.status(400).send('invalid from date given');
+    }
+
+    if (Number.isNaN(to)) {
+      return res.status(400).send('invalid to date given');
+    }
+
+    const type = req.params.type;
+
+    if (!this.transactions.hasIdentifier(type)) {
+      return res.status(400).send('invalid type given');
+    }
+
+    if (to <= from || to - from > 100) {
+      return res.status(400).send('invalid period range given');
+    }
+
+    const stats = await this.transactions.getStats(type, from, to);
+
+    res.status(200).json(stats);
   }
 }
