+++ src/transaction/transaction.controller.spec.ts
@@ -3,12 +3,10 @@ import { Test, TestingModule } from '@nestjs/testing';
 import request from 'supertest';
 import { TransactionModuleConfig } from './transaction.module';
 import { NodeService } from '../node/node.service';
-import { TransactionService } from './transaction.service';
 
 describe('TransactionController', () => {
   let module: TestingModule;
   let nodeService: NodeService;
-  let txService: TransactionService;
   let app: INestApplication;
 
   function spy() {
@@ -22,16 +20,7 @@ describe('TransactionController', () => {
         .mockImplementation(async () => [{ id: 'fake_transaction_1' }, { id: 'fake_transaction_2' }]),
     };
 
-    const tx = {
-      getStats: jest.spyOn(txService, 'getStats')
-        .mockImplementation(async () => [
-          { period: '2021-03-01 00:00:00', count: 56847 },
-          { period: '2021-03-02 00:00:00', count: 103698 },
-          { period: '2021-03-03 00:00:00', count: 33329 },
-        ]),
-    };
-
-    return { node, tx };
+    return { node };
   }
 
   beforeEach(async () => {
@@ -40,7 +29,6 @@ describe('TransactionController', () => {
     await app.init();
 
     nodeService = module.get<NodeService>(NodeService);
-    txService = module.get<TransactionService>(TransactionService);
   });
 
   afterEach(async () => {
@@ -101,50 +89,6 @@ describe('TransactionController', () => {
 
       expect(spies.node.getTransactions.mock.calls.length).toBe(1);
       expect(spies.node.getTransactions.mock.calls[0][0]).toEqual(['fake_transaction_1', 'fake_transaction_2']);
-    });
-  });
-
-  describe('GET /transactions/stats/:type/:from/:to', () => {
-    test('should get stats using timestamps', async () => {
-      const spies = spy();
-
-      const res = await request(app.getHttpServer())
-        .get(`/transactions/stats/all/1614597904336/1614793341900`)
-        .send();
-
-      expect(res.status).toBe(200);
-      expect(res.header['content-type']).toBe('application/json; charset=utf-8');
-      expect(res.body).toEqual([
-        { period: '2021-03-01 00:00:00', count: 56847 },
-        { period: '2021-03-02 00:00:00', count: 103698 },
-        { period: '2021-03-03 00:00:00', count: 33329 },
-      ]);
-
-      expect(spies.tx.getStats.mock.calls.length).toBe(1);
-      expect(spies.tx.getStats.mock.calls[0][0]).toBe('all');
-      expect(spies.tx.getStats.mock.calls[0][1]).toBe(18687);
-      expect(spies.tx.getStats.mock.calls[0][2]).toBe(18689);
-    });
-
-    test('should get stats using date strings', async () => {
-      const spies = spy();
-
-      const res = await request(app.getHttpServer())
-        .get(`/transactions/stats/all/2021-03-01/2021-03-03`)
-        .send();
-
-      expect(res.status).toBe(200);
-      expect(res.header['content-type']).toBe('application/json; charset=utf-8');
-      expect(res.body).toEqual([
-        { period: '2021-03-01 00:00:00', count: 56847 },
-        { period: '2021-03-02 00:00:00', count: 103698 },
-        { period: '2021-03-03 00:00:00', count: 33329 },
-      ]);
-
-      expect(spies.tx.getStats.mock.calls.length).toBe(1);
-      expect(spies.tx.getStats.mock.calls[0][0]).toBe('all');
-      expect(spies.tx.getStats.mock.calls[0][1]).toBe(18687);
-      expect(spies.tx.getStats.mock.calls[0][2]).toBe(18689);
     });
   });
 });
