+++ src/transaction/transaction.controller.ts
@@ -77,46 +77,4 @@ export class TransactionController {
       return res.status(500).send(`failed to get transaction by address '${e}'`);
     }
   }
-
-  @Get('stats/:type/:from/:to')
-  @ApiOperation({ summary: 'Get transaction count per day' })
-  @ApiParam({ name: 'type', description: 'Transaction type'})
-  @ApiParam({ name: 'from', description: 'Date as `year-month-day` or timestamp in ms' })
-  @ApiParam({ name: 'to', description: 'Date as `year-month-day` or timestamp in ms' })
-  @ApiResponse({ status: 200 })
-  @ApiResponse({
-    status: 400,
-    description: ['invalid from date given', 'invalid to date given', 'invalid type given',
-      'invalid period range given'].join('<br>'),
-  })
-  @ApiResponse({ status: 500, description: `failed to get transaction stats '[reason]'` })
-  async getTransactionStats(@Req() req: Request, @Res() res: Response): Promise<Response> {
-    const fromParam = req.params.from;
-    const toParam = req.params.to;
-
-    let from = 0;
-    let to = 0;
-
-    from = Math.floor(new Date(fromParam.match(/\D/) ? fromParam : Number(fromParam)).getTime() / 86400000);
-    to = Math.floor(new Date(toParam.match(/\D/) ? toParam : Number(toParam)).getTime() / 86400000);
-    if (Number.isNaN(from)) {
-      return res.status(400).send('invalid from date given');
-    }
-    if (Number.isNaN(to)) {
-      return res.status(400).send('invalid to date given');
-    }
-
-    const type = req.params.type as txType;
-    if (!this.tx.hasIdentifier(type)) {
-      return res.status(400).send('invalid type given');
-    }
-
-    if (to <= from || to - from > 100) {
-      return res.status(400).send('invalid period range given');
-    }
-
-    const stats = await this.tx.getStats(type, from, to);
-
-    res.status(200).json(stats);
-  }
 }
